requirements.txt
llama-index-vector-stores-qdrant>=0.6.0
llama-index>=0.12.23
fastapi>=0.115.11
crewai>=0.105.0
crewai-tools>=0.38.0
kafka-python>=2.0.6
python-dotenv==1.0.1
prometheus_client>=0.21.1
httpx>=0.27.2
uvicorn>=0.34.0
pydantic>=2.10.6
qdrant-client>=1.13.3
backoff>=2.2.1
llama-index-core>=0.12.41
prometheus_api_client>=0.6.0
